+++ postgres/src/lib.rs
@@ -97,7 +97,6 @@ pub mod compat {
     //!
     //! [`futures::Stream`]: futures_core::stream::Stream
 
-    pub use crate::row::compat::RowOwned;
     pub use crate::statement::compat::StatementGuarded;
 }