+++ postgres/src/row.rs
@@ -3,7 +3,4 @@
 mod traits;
 mod types;
 
-#[cfg(feature = "compat")]
-pub(crate) mod compat;
-
-pub use types::{Row, RowSimple};
+pub use types::{Row, RowOwned, RowSimple};