+++ postgres/CHANGES.md
@@ -32,6 +32,8 @@
 - `Encode` and `IntoStream` traits implementation detail change
 
 ## Add
+- add `Client::prepare_blocking`
+- add `Prepare::{_prepare_blocking, _get_type_blocking}`
 - add `iter::AsyncLendingIteratorExt` for extending async iterator APIs
 - add `statement::Statement::{bind, bind_dyn}` methods for binding value parameters to a prepared statement for query
 - add `error::DriverIoErrorMulti` type for outputting read and write IO errors at the same time